+++ lib/NodeType.js
