+++ lib/TokenType.js
