+++ lib/TypeDictionary.js
@@ -1,414 +1,245 @@
-// This script licensed under the MIT.
-// http://orgachem.mit-license.org
-
-
-var util = require('./util.js');
-
-
+'use strict';
 
 /**
- * A singleton class for type dictionaries.  This dictionaty is useful to check
- * an user define type name.  Maybe, a type name publisher needs the check to
- * determine whether should convert a type name to a link.
- *
- * Use {@code getInstance()} to get the instance.
- *
- * For example:
- * <pre>
- * var dict = TypeDictionary.getInstance();
- *
- * // Browser objects
- * dict.has('string'); // ⇒ true
- * dict.has('Element'); // ⇒ true
- *
- * // User define objects
- * dict.has('FooBar'); // ⇒ false
- * dict.has('foo.Bar'); // ⇒ false
- * </pre>
- * @constructor
- * @exports lib/TypeDictionary
+ * Namespace for type dictionaries.
+ * @type {Object<string, boolean>}
  */
-var TypeDictionary = function() {
-  var dict = [].concat(TypeDictionary.GENERALS);
+var TypeDictionary = {
+  'Array': true,
+  'Boolean': true,
+  'Date': true,
+  'Document': true,
+  'Element': true,
+  'Error': true,
+  'Function': true,
+  'Image': true,
+  'Infinity': true,
+  'Iterator': true,
+  'NaN': true,
+  'Null': true,
+  'Number': true,
+  'Object': true,
+  'Option': true,
+  'RegExp': true,
+  'String': true,
+  'Undefined': true,
+  'Window': true,
 
-  if (TypeDictionary.ERROR_TYPES_ENABLED) {
-    dict.push.apply(dict, TypeDictionary.ERRORS);
-  }
+  'EvalError': true,
+  'InternalError': true,
+  'RangeError': true,
+  'ReferenceError': true,
+  'StopIteration': true,
+  'SyntaxError': true,
+  'TypeError': true,
+  'UriError': true,
 
-  if (TypeDictionary.HTML_TYPES_ENABLED) {
-    dict.push.apply(dict, TypeDictionary.HTMLS);
-  }
+  'Attr': true,
+  'CDATASection': true,
+  'CharacterData': true,
+  'Comment': true,
+  'console': true,
+  'DocumentFragment': true,
+  'DocumentType': true,
+  'DomConfiguration': true,
+  'DOMError': true,
+  'DOMErrorHandler': true,
+  'DOMException': true,
+  'DOMImplementation': true,
+  'DOMImplementationList': true,
+  'DOMImplementationRegistry': true,
+  'DOMImplementationSource': true,
+  'DOMLocator': true,
+  'DOMObject': true,
+  'DOMRequest': true,
+  'DOMString': true,
+  'DOMStringList': true,
+  'DOMTimeStamp': true,
+  'DOMUserData': true,
+  'Entity': true,
+  'EntityReference': true,
+  'MediaQueryList': true,
+  'MediaQueryListListener': true,
+  'NameList': true,
+  'NamedNodeMap': true,
+  'Node': true,
+  'NodeFilter': true,
+  'NodeIterator': true,
+  'NodeList': true,
+  'Notation': true,
+  'Plugin': true,
+  'PluginArray': true,
+  'ProcessingInstruction': true,
+  'SharedWorker': true,
+  'Text': true,
+  'TimeRanges': true,
+  'Treewalker': true,
+  'TypeInfo': true,
+  'UserDataHandler': true,
+  'Worker': true,
+  'WorkerGlobalScope': true,
 
-  if (TypeDictionary.DOM_TYPES_ENABLED) {
-    dict.push.apply(dict, TypeDictionary.DOMS);
-  }
+  'HTMLDocument': true,
+  'HTMLElement': true,
+  'HTMLAnchorElement': true,
+  'HTMLAppletElement': true,
+  'HTMLAudioElement': true,
+  'HTMLAreaElement': true,
+  'HTMLBaseElement': true,
+  'HTMLBaseFontElement': true,
+  'HTMLBodyElement': true,
+  'HTMLBRElement': true,
+  'HTMLButtonElement': true,
+  'HTMLCanvasElement': true,
+  'HTMLDataListElement': true,
+  'HTMLDirectoryElement': true,
+  'HTMLDivElement': true,
+  'HTMLDListElement': true,
+  'HTMLEmbedElement': true,
+  'HTMLFieldSetElement': true,
+  'HTMLFontElement': true,
+  'HTMLFormElement': true,
+  'HTMLFrameElement': true,
+  'HTMLFrameSetElement': true,
+  'HTMLHeadElement': true,
+  'HTMLHeadingElement': true,
+  'HTMLHtmlElement': true,
+  'HTMLHRElement': true,
+  'HTMLIFrameElement': true,
+  'HTMLImageElement': true,
+  'HTMLInputElement': true,
+  'HTMLKeygenElement': true,
+  'HTMLLabelElement': true,
+  'HTMLLIElement': true,
+  'HTMLLinkElement': true,
+  'HTMLMapElement': true,
+  'HTMLMenuElement': true,
+  'HTMLMetaElement': true,
+  'HTMLMeterElement': true,
+  'HTMLModElement': true,
+  'HTMLObjectElement': true,
+  'HTMLOListElement': true,
+  'HTMLOptGroupElement': true,
+  'HTMLOptionElement': true,
+  'HTMLOutputElement': true,
+  'HTMLParagraphElement': true,
+  'HTMLParamElement': true,
+  'HTMLPreElement': true,
+  'HTMLProgressElement': true,
+  'HTMLQuoteElement': true,
+  'HTMLScriptElement': true,
+  'HTMLSelectElement': true,
+  'HTMLSourceElement': true,
+  'HTMLSpanElement': true,
+  'HTMLStyleElement': true,
+  'HTMLTableElement': true,
+  'HTMLTableCaptionElement': true,
+  'HTMLTableCellElement: true,': true,
+  'HTMLTableDataCellElement': true,
+  'HTMLTableHeaderCellElement': true,
+  'HTMLTableColElement': true,
+  'HTMLTableRowElement': true,
+  'HTMLTableSectionElement': true,
+  'HTMLTextAreaElement': true,
+  'HTMLTimeElement': true,
+  'HTMLTitleElement': true,
+  'HTMLTrackElement': true,
+  'HTMLUListElement': true,
+  'HTMLUnknownElement': true,
+  'HTMLVideoElement': true,
+  'CanvasRenderingContext2D': true,
+  'CanvasGradient': true,
+  'CanvasPattern': true,
+  'TextMetrics': true,
+  'ImageData': true,
+  'CanvasPixelArray': true,
 
-  if (TypeDictionary.SVGS_TYPES_ENABLED) {
-    dict.push.apply(dict, TypeDictionary.SVGS);
-  }
-
-  dict = dict.map(function(word) {
-    return word.toLowerCase();
-  });
-
-  dict.sort();
-
-  this.dict_ = dict;
-  this.dictLen_ = dict.length;
+  'SVGElement': true,
+  'SVGAElement': true,
+  'SVGAltGlyphElement': true,
+  'SVGAltGlyphDefElement': true,
+  'SVGAltGlyphItemElement': true,
+  'SVGAnimationElement': true,
+  'SVGAnimateElement': true,
+  'SVGAnimateColorElement': true,
+  'SVGAnimateMotionElement': true,
+  'SVGAnimateTransformElement': true,
+  'SVGSetElement': true,
+  'SVGCircleElement': true,
+  'SVGClipPathElement': true,
+  'SVGColorProfileElement': true,
+  'SVGCursorElement': true,
+  'SVGDefsElement': true,
+  'SVGDescElement': true,
+  'SVGEllipseElement': true,
+  'SVGFilterElement': true,
+  'SVGFilterPrimitiveStandardAttributes': true,
+  'SVGFEBlendElement': true,
+  'SVGFEColorMatrixElement': true,
+  'SVGFEComponentTransferElement': true,
+  'SVGFECompositeElement': true,
+  'SVGFEConvolveMatrixElement': true,
+  'SVGFEDiffuseLightingElement': true,
+  'SVGFEDisplacementMapElement': true,
+  'SVGFEDistantLightElement': true,
+  'SVGFEFloodElement': true,
+  'SVGFEGaussianBlurElement': true,
+  'SVGFEImageElement': true,
+  'SVGFEMergeElement': true,
+  'SVGFEMergeNodeElement': true,
+  'SVGFEMorphologyElement': true,
+  'SVGFEOffsetElement': true,
+  'SVGFEPointLightElement': true,
+  'SVGFESpecularLightingElement': true,
+  'SVGFESpotLightElement': true,
+  'SVGFETileElement': true,
+  'SVGFETurbulenceElement': true,
+  'SVGComponentTransferFunctionElement': true,
+  'SVGFEFuncRElement': true,
+  'SVGFEFuncGElement': true,
+  'SVGFEFuncBElement': true,
+  'SVGFEFuncAElement': true,
+  'SVGFontElement': true,
+  'SVGFontFaceElement': true,
+  'SVGFontFaceFormatElement': true,
+  'SVGFontFaceNameElement': true,
+  'SVGFontFaceSrcElement': true,
+  'SVGFontFaceUriElement': true,
+  'SVGForeignObjectElement': true,
+  'SVGGElement': true,
+  'SVGGlyphElement': true,
+  'SVGGlyphRefElement': true,
+  'SVGGradientElement': true,
+  'SVGLinearGradientElement': true,
+  'SVGRadialGradientElement': true,
+  'SVGHKernElement': true,
+  'SVGImageElement': true,
+  'SVGLineElement': true,
+  'SVGMarkerElement': true,
+  'SVGMaskElement': true,
+  'SVGMetadataElement': true,
+  'SVGMissingGlyphElement': true,
+  'SVGMPathElement': true,
+  'SVGPathElement': true,
+  'SVGPatternElement': true,
+  'SVGPolylineElement': true,
+  'SVGPolygonElement': true,
+  'SVGRectElement': true,
+  'SVGScriptElement': true,
+  'SVGStopElement': true,
+  'SVGStyleElement': true,
+  'SVGSVGElement': true,
+  'SVGSwitchElement': true,
+  'SVGSymbolElement': true,
+  'SVGTextElement': true,
+  'SVGTextPathElement': true,
+  'SVGTitleElement': true,
+  'SVGTRefElement': true,
+  'SVGTSpanElement': true,
+  'SVGUseElement': true,
+  'SVGViewElement': true,
+  'SVGVKernElement': true
 };
-util.addSingletonGetter(TypeDictionary);
-
-
-/**
- * Whether error object types are enabled.  Default is true.
- * @const
- * @type {boolean}
- */
-TypeDictionary.ERROR_TYPES_ENABLED = true;
-
-
-/**
- * Whether DOM interface types are enabled.  Default is true.
- * @const
- * @type {boolean}
- */
-TypeDictionary.DOM_TYPES_ENABLED = true;
-
-
-/**
- * Whether HTML interface types are enabled.  Default is true.
- * @const
- * @type {boolean}
- */
-TypeDictionary.HTML_TYPES_ENABLED = true;
-
-
-/**
- * Whether SVG interface types are enabled.  Default is true.
- * @const
- * @type {boolean}
- */
-TypeDictionary.SVGS_TYPES_ENABLED = true;
-
-
-/**
- * Type word array.  The array has to be sorted.
- * @type {Array.<string>}
- * @private
- */
-TypeDictionary.prototype.dict_ = null;
-
-
-/**
- * Type word array length.
- * @type {number}
- * @private
- */
-TypeDictionary.prototype.dictLen_ = null;
-
-
-/**
- * Whether the specified type is in the dictionary.
- * @param {string} target Type to check.
- * @return {boolean} Whether the specified type is in the dictionary.
- */
-TypeDictionary.prototype.has = function(target) {
-  var a = target.toLowerCase();
-  var arr = this.dict_;
-  var left = 0;  // inclusive
-  var right = this.dictLen_;  // exclusive
-  var found;
-  while (left < right) {
-    var middle = (left + right) >> 1;
-    var b = arr[middle];
-    var compareResult = a > b ? 1 : a < b ? -1 : 0;
-    if (compareResult > 0) {
-      left = middle + 1;
-    } else if (compareResult === 0) {
-      return true;
-    } else {
-      right = middle;
-    }
-  }
-  return false;
-};
-
-
-/**
- * General purpose object names.
- * @const
- * @type {Array.<string>}
- */
-TypeDictionary.GENERALS = [
-  'Array',
-  'Boolean',
-  'Date',
-  'Document',
-  'Element',
-  'Error',
-  'Function',
-  'Image',
-  'Infinity',
-  'Iterator',
-  'NaN',
-  'Null',
-  'Number',
-  'Object',
-  'Option',
-  'RegExp',
-  'String',
-  'Undefined',
-  'Window',
-  '*', // All type
-  '?' // Unkown type
-];
-
-
-/**
- * Error object names.
- * @const
- * @type {Array.<string>}
- */
-TypeDictionary.ERRORS = [
-  'EvalError',
-  'InternalError',
-  'RangeError',
-  'ReferenceError',
-  'StopIteration',
-  'SyntaxError',
-  'TypeError',
-  'UriError'
-];
-
-
-/**
- * DOM object names.
- * @const
- * @type {Array.<string>}
- */
-TypeDictionary.DOMS = [
-  'Attr',
-  'CDATASection',
-  'CharacterData',
-  'Comment',
-  'console',
-  'DocumentFragment',
-  'DocumentType',
-  'DomConfiguration',
-  'DOMError',
-  'DOMErrorHandler',
-  'DOMException',
-  'DOMImplementation',
-  'DOMImplementationList',
-  'DOMImplementationRegistry',
-  'DOMImplementationSource',
-  'DOMLocator',
-  'DOMObject',
-  'DOMRequest',
-  'DOMString',
-  'DOMStringList',
-  'DOMTimeStamp',
-  'DOMUserData',
-  'Entity',
-  'EntityReference',
-  'MediaQueryList',
-  'MediaQueryListListener',
-  'NameList',
-  'NamedNodeMap',
-  'Node',
-  'NodeFilter',
-  'NodeIterator',
-  'NodeList',
-  'Notation',
-  'Plugin',
-  'PluginArray',
-  'ProcessingInstruction',
-  'SharedWorker',
-  'Text',
-  'TimeRanges',
-  'Treewalker',
-  'TypeInfo',
-  'UserDataHandler',
-  'Worker',
-  'WorkerGlobalScope'
-];
-
-
-/**
- * HTML interface names.
- * @const
- * @type {Array.<string>}
- */
-TypeDictionary.HTMLS = [
-  'HTMLDocument',
-  'HTMLElement',
-  'HTMLAnchorElement',
-  'HTMLAppletElement',
-  'HTMLAudioElement',
-  'HTMLAreaElement',
-  'HTMLBaseElement',
-  'HTMLBaseFontElement',
-  'HTMLBodyElement',
-  'HTMLBRElement',
-  'HTMLButtonElement',
-  'HTMLCanvasElement',
-  'HTMLDataListElement',
-  'HTMLDirectoryElement',
-  'HTMLDivElement',
-  'HTMLDListElement',
-  'HTMLEmbedElement',
-  'HTMLFieldSetElement',
-  'HTMLFontElement',
-  'HTMLFormElement',
-  'HTMLFrameElement',
-  'HTMLFrameSetElement',
-  'HTMLHeadElement',
-  'HTMLHeadingElement',
-  'HTMLHtmlElement',
-  'HTMLHRElement',
-  'HTMLIFrameElement',
-  'HTMLImageElement',
-  'HTMLInputElement',
-  'HTMLKeygenElement',
-  'HTMLLabelElement',
-  'HTMLLIElement',
-  'HTMLLinkElement',
-  'HTMLMapElement',
-  'HTMLMenuElement',
-  'HTMLMetaElement',
-  'HTMLMeterElement',
-  'HTMLModElement',
-  'HTMLObjectElement',
-  'HTMLOListElement',
-  'HTMLOptGroupElement',
-  'HTMLOptionElement',
-  'HTMLOutputElement',
-  'HTMLParagraphElement',
-  'HTMLParamElement',
-  'HTMLPreElement',
-  'HTMLProgressElement',
-  'HTMLQuoteElement',
-  'HTMLScriptElement',
-  'HTMLSelectElement',
-  'HTMLSourceElement',
-  'HTMLSpanElement',
-  'HTMLStyleElement',
-  'HTMLTableElement',
-  'HTMLTableCaptionElement',
-  'HTMLTableCellElement,',
-  'HTMLTableDataCellElement',
-  'HTMLTableHeaderCellElement',
-  'HTMLTableColElement',
-  'HTMLTableRowElement',
-  'HTMLTableSectionElement',
-  'HTMLTextAreaElement',
-  'HTMLTimeElement',
-  'HTMLTitleElement',
-  'HTMLTrackElement',
-  'HTMLUListElement',
-  'HTMLUnknownElement',
-  'HTMLVideoElement',
-  'CanvasRenderingContext2D',
-  'CanvasGradient',
-  'CanvasPattern',
-  'TextMetrics',
-  'ImageData',
-  'CanvasPixelArray'
-];
-
-
-/**
- * SVG interface names.
- * @const
- * @type {Array.<string>}
- */
-TypeDictionary.SVGS = [
-  'SVGElement',
-  'SVGAElement',
-  'SVGAltGlyphElement',
-  'SVGAltGlyphDefElement',
-  'SVGAltGlyphItemElement',
-  'SVGAnimationElement',
-  'SVGAnimateElement',
-  'SVGAnimateColorElement',
-  'SVGAnimateMotionElement',
-  'SVGAnimateTransformElement',
-  'SVGSetElement',
-  'SVGCircleElement',
-  'SVGClipPathElement',
-  'SVGColorProfileElement',
-  'SVGCursorElement',
-  'SVGDefsElement',
-  'SVGDescElement',
-  'SVGEllipseElement',
-  'SVGFilterElement',
-  'SVGFilterPrimitiveStandardAttributes',
-  'SVGFEBlendElement',
-  'SVGFEColorMatrixElement',
-  'SVGFEComponentTransferElement',
-  'SVGFECompositeElement',
-  'SVGFEConvolveMatrixElement',
-  'SVGFEDiffuseLightingElement',
-  'SVGFEDisplacementMapElement',
-  'SVGFEDistantLightElement',
-  'SVGFEFloodElement',
-  'SVGFEGaussianBlurElement',
-  'SVGFEImageElement',
-  'SVGFEMergeElement',
-  'SVGFEMergeNodeElement',
-  'SVGFEMorphologyElement',
-  'SVGFEOffsetElement',
-  'SVGFEPointLightElement',
-  'SVGFESpecularLightingElement',
-  'SVGFESpotLightElement',
-  'SVGFETileElement',
-  'SVGFETurbulenceElement',
-  'SVGComponentTransferFunctionElement',
-  'SVGFEFuncRElement',
-  'SVGFEFuncGElement',
-  'SVGFEFuncBElement',
-  'SVGFEFuncAElement',
-  'SVGFontElement',
-  'SVGFontFaceElement',
-  'SVGFontFaceFormatElement',
-  'SVGFontFaceNameElement',
-  'SVGFontFaceSrcElement',
-  'SVGFontFaceUriElement',
-  'SVGForeignObjectElement',
-  'SVGGElement',
-  'SVGGlyphElement',
-  'SVGGlyphRefElement',
-  'SVGGradientElement',
-  'SVGLinearGradientElement',
-  'SVGRadialGradientElement',
-  'SVGHKernElement',
-  'SVGImageElement',
-  'SVGLineElement',
-  'SVGMarkerElement',
-  'SVGMaskElement',
-  'SVGMetadataElement',
-  'SVGMissingGlyphElement',
-  'SVGMPathElement',
-  'SVGPathElement',
-  'SVGPatternElement',
-  'SVGPolylineElement',
-  'SVGPolygonElement',
-  'SVGRectElement',
-  'SVGScriptElement',
-  'SVGStopElement',
-  'SVGStyleElement',
-  'SVGSVGElement',
-  'SVGSwitchElement',
-  'SVGSymbolElement',
-  'SVGTextElement',
-  'SVGTextPathElement',
-  'SVGTitleElement',
-  'SVGTRefElement',
-  'SVGTSpanElement',
-  'SVGUseElement',
-  'SVGViewElement',
-  'SVGVKernElement'
-];
 
 
 // Exports the constructor.
