--- test/run.js
@@ -1,19 +1,0 @@
-#!/usr/bin/env node
-// This script licensed under the MIT.
-// http://orgachem.mit-license.org
-
-try {
-    var reporter = require('nodeunit').reporters.default;
-}
-catch(e) {
-    console.log("Cannot find nodeunit module.");
-    console.log("You can download submodules for this project by doing:");
-    console.log("");
-    console.log("    git submodule init");
-    console.log("    git submodule update");
-    console.log("");
-    process.exit();
-}
-
-process.chdir(__dirname);
-reporter.run(['tests']);
